backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
mysql-connector-python
python-dotenv
gunicorn
pymysql
cryptography
flask-jwt-extended